Now only predicting binary 'up or down'
I don't expect this to have much success, but I had to try.
With the last commit I have achieved accuracy of about 53%, which is
still quite low. Also, because of line 160 in nn.js (because of those
equalities) this might even mean accuracy of less than 50%. Not sure. I
will consider this project a success if I achieve more than 60%, or at
least 55% if I'm sure the accuracy is calculated properly.

After testing this commit: training just for several minutes and testing
with 10000 charts, these are the accuracies (up or down accuracies):
46.12% (horrible)
54.42% (lucky?)
46.72%
I might try and train it for several hours, but I'm not expecting it to
get accurate at this point.