Probably the last commit
Added license.
Finished (probably) the README.
At this point, the AI takes 50 historical candles (prices and volumes),
EMA and RSI indicators. Based on that, it predicts the change of the
next candle relative to the last known (historical) candle in %.
Check the README for the accuracy.

At this point, I no longer believe in this project and I'm abandoning
it. I might get back to it but I don't think I will. I will also
probably take this down of the internet and make it private.